Refactoring ctypes module to a PyPy-like file tree
